pyproject.toml
[tool.poetry]
name = "rubbi-official-bot-by-397"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nextcord = "^2.2.0"
motor = "^3.0.0"
asyncpg = "^0.26.0"
Pillow = "^9.3.0"
requests = "^2.28.1"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
